Macro refactoring: Add a CLOSy API to describe commands and parameters
Overview:

Encoding and decoding is done using class PARAMETER-INFO, which describes
each value in the protocol using its name, type (and, if required, special
configuration like the length of the data).

Class FRAME-INFO describes a protocol frame, i.e. a list of parameters,
with a function to compute the length from that, and functions to read
and write frames.

The named FRAME-PAIR instances describe request and reply in
combination.  Macros defining frames fill in the FRAME-PAIR object.